+++ config/services.php
@@ -36,6 +36,7 @@ return [
         'secret' => env('STRIPE_SECRET', null),
         'webhook' => [
             'secret' => env('STRIPE_WEBHOOK_SECRET', null),
+            'tolerance' => env('STRIPE_WEBHOOK_TOLERANCE', 300),
         ],
     ],